+++ src/main/java/nextstep/domain/reservation/ReservationService.java
@@ -5,8 +5,8 @@ import nextstep.domain.schedule.ScheduleDao;
 import nextstep.domain.member.MemberDao;
 import nextstep.domain.member.Member;
 import nextstep.interfaces.reservation.dto.ReservationRequest;
-import nextstep.support.DuplicateEntityException;
-import nextstep.support.NotExistEntityException;
+import nextstep.interfaces.DuplicateEntityException;
+import nextstep.interfaces.NotExistEntityException;
 import nextstep.domain.theme.Theme;
 import nextstep.domain.theme.ThemeDao;
 import org.springframework.stereotype.Service;
