+++ src/main/java/nextstep/interfaces/DuplicateEntityException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.interfaces;
 
 public class DuplicateEntityException extends RuntimeException {
 }
