+++ src/main/java/nextstep/interfaces/NotExistEntityException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.interfaces;
 
 public class NotExistEntityException extends RuntimeException {
 }
